pkg/endpoint: move UpdateLabels to labels.go
Signed-off by: Ian Vernon <ian@cilium.io>
